We follow the AffiliateWP’s business model and open the development to everyone willing to contribute! Read the blog post which convinced us:  http://wptavern.com/affiliatewp-passes-30k-in-monthly-revenue-after-2-years-in-business
We believe that our business model of a commercial core plugin and free extensions is the future as it gives you freedom and no hidden costs in every new functionality / extension you need ;)

Let us find out how it works together! We have a great feeling and welcome you all to contribute and become a part of the BuddyForms Team.