Remove edit attachment link from wp media upload if used in frontend forms Add new filter to change the post per page in the user post list: bf_user_posts_query_args_posts_per_page
Fix a issue with the slug met box. We like to have it hidden as we not want to change the slug. The plugin heavily depends on the slug at the moment.
We should switch from slug to id based logic. Slug is left over from the time where no post type was used to save options. I will leave this to a separate update. It requires database changes and needs to get tested solid before it gets released.