Fixed a complex issue with attachments. Thanks for Farnousch pointing this out so clear. Still not 100%
If a featured image or any file is uploaded via the media uploader, it gets attached to the page it gest uploaded from and not to the post you are about to create. This is logical because in time when you upload the file the post is not created.

Every attachment can only have one parent_post. This parent_post is used as "uploaded to"

If you use get_attached_media() you only get the media uploaded to the post and not the once you use in the post but got uploaded earlier and just reused.

If I upload two images to a post this post will have the two images in the callback of get_attached_media() If I reuse this same images in a other post get_attached_media() will return an empty array.

I try to follow the exact same logic. During post creation we check if the attachments parent post is == the form attached post. And this is the trouble. This only works if the attached page is used to display the post.

I have not found a better solution till now. Also I not understand the sense of get_attached_media() limit to the once uploaded with the post.