Huge rewrite of the validation process and error handling. All errors should get handled with the same error class.
Switch form submission to the pfbc and rewrite the pfbc to work with WP.

Rename editpost_ to buddyforms_form_ to keep it more easy readable.

Split submissions-single into two templates one for the frontend people are able to overwrite in there themes and one for the admin submission single view with more user data.